feat: Add no results for filter message to My Collection
