+++ src/lib/Scenes/MyCollection/MyCollectionArtworks.tsx
@@ -1,0 +1,92 @@
+import { addCollectedArtwork } from "@artsy/cohesion"
+import { MyCollection_me } from "__generated__/MyCollection_me.graphql"
+import { ArtworksFiltersStore } from "lib/Components/ArtworkFilter/ArtworkFilterStore"
+import { FilteredArtworkGridZeroState } from "lib/Components/ArtworkGrids/FilteredArtworkGridZeroState"
+import { InfiniteScrollMyCollectionArtworksGridContainer } from "lib/Components/ArtworkGrids/InfiniteScrollArtworksGrid"
+import { ZeroState } from "lib/Components/States/ZeroState"
+import { navigate, popToRoot } from "lib/navigation/navigate"
+import { extractNodes } from "lib/utils/extractNodes"
+import { Button, Flex } from "palette"
+import React from "react"
+import { RelayPaginationProp } from "react-relay"
+import { useTracking } from "react-tracking"
+import { MyCollectionArtworkEdge } from "./MyCollection"
+import { localSortAndFilterArtworks } from "./utils/localArtworkSortAndFilter"
+
+interface MyCollectionArtworksProps {
+  me: MyCollection_me
+  keywordFilter: string
+  relay: RelayPaginationProp
+}
+
+export const MyCollectionArtworks: React.FC<MyCollectionArtworksProps> = ({
+  me,
+  keywordFilter,
+  relay,
+}) => {
+  const appliedFiltersState = ArtworksFiltersStore.useStoreState((state) => state.appliedFilters)
+  const filterOptions = ArtworksFiltersStore.useStoreState((state) => state.filterOptions)
+
+  const artworks = extractNodes(me?.myCollectionConnection)
+
+  const filteredArtworks = localSortAndFilterArtworks(
+    artworks as any,
+    appliedFiltersState,
+    filterOptions,
+    keywordFilter
+  )
+
+  if (artworks.length === 0) {
+    return <MyCollectionZeroState />
+  }
+
+  if (filteredArtworks.length === 0) {
+    return (
+      <Flex py="6" px="2">
+        <FilteredArtworkGridZeroState hideClearButton />
+      </Flex>
+    )
+  }
+
+  return (
+    <Flex pt={1}>
+      <InfiniteScrollMyCollectionArtworksGridContainer
+        myCollectionConnection={me.myCollectionConnection!}
+        hasMore={relay.hasMore}
+        loadMore={relay.loadMore}
+        // tslint:disable-next-line: no-shadowed-variable
+        localSortAndFilterArtworks={(artworks: MyCollectionArtworkEdge[]) =>
+          localSortAndFilterArtworks(artworks, appliedFiltersState, filterOptions, keywordFilter)
+        }
+      />
+    </Flex>
+  )
+}
+
+const MyCollectionZeroState: React.FC = () => {
+  const { trackEvent } = useTracking()
+
+  return (
+    <ZeroState
+      title="Your art collection in your pocket."
+      subtitle="Keep track of your collection all in one place and get market insights"
+      callToAction={
+        <Button
+          testID="add-artwork-button-zero-state"
+          onPress={() => {
+            trackEvent(addCollectedArtwork())
+            navigate("my-collection/artworks/new", {
+              passProps: {
+                mode: "add",
+                onSuccess: popToRoot,
+              },
+            })
+          }}
+          block
+        >
+          Add artwork
+        </Button>
+      }
+    />
+  )
+}
